use CSS for drawing area blur effect instead
use CSS to implment the background color and drawing area blur effect.
we don't need to modify paper.js for blur effects anymore so change it
back to offical version.

the performance of showing the blurred drawing area is also greatly
improved when loading the webpage. Now we don't see the latency effect.