now the solver hooks up with dimension UI
For the first time the solver works with the dimension UI.  We can
define a distance-point-line constraint between a rectangle's edge and
an oval's center, then modify its offset value. The solver could put the
elements to positions that satisify the constraint.

Kudos! there're still some bugs, I'll look after them in the coming
days.